Move event_bus.py into RimSort.utils
