+++ RimSort/controllers/settings_controller.py
@@ -9,7 +9,7 @@ from PySide6.QtWidgets import QFileDialog, QMessageBox, QApplication
 from loguru import logger
 
 from RimSort.models.settings import Settings
-from util.event_bus import EventBus
+from RimSort.utils.event_bus import EventBus
 from util.generic import platform_specific_open
 from RimSort.utils.system_info import SystemInfo
 from view.settings_dialog import SettingsDialog
