+++ model/mod_list.py
@@ -20,7 +20,7 @@ from PySide6.QtWidgets import (
 from RimSort.controllers.settings_controller import SettingsController
 from model.mod_list_item import ModListItemInner
 from model.dialogue import show_dialogue_conditional, show_dialogue_input, show_warning
-from util.event_bus import EventBus
+from RimSort.utils.event_bus import EventBus
 from util.generic import (
     delete_files_except_extension,
     handle_remove_read_only,
